feat: send spring csrf token for endpoint calls
